+++ src/core/orchestrator/index.js
@@ -10,7 +10,7 @@ const logger = createLogger('OmnySys:orchestrator');
 
 import * as lifecycle from './lifecycle/index.js';
 import * as queueing from './queueing.js';
-import * as llmAnalysis from './llm-analysis.js';
+import * as staticInsights from './static-insights.js';
 import * as iterative from './iterative.js';
 import * as issues from './issues.js';
 import * as helpers from './helpers.js';
@@ -196,7 +196,7 @@ Object.assign(
   Orchestrator.prototype,
   lifecycle,
   queueing,
-  llmAnalysis,
+  staticInsights,
   iterative,
   issues,
   helpers